restructures files and makes directories for Routes and Components
