+++ src/Routes/SavedRoutines.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import '../App.css';
 import {
     Link,
     Route,
